+++ src/packages/common/component-handler/init-functional-dialog.js
@@ -45,5 +45,9 @@ export function initFunctionalDialog({
     return promise.then(val => Promise.resolve(val))
       .catch(err => Promise.reject(err));
   }
-  console.warn(`The ${options.selector} node is not found, please confirm whether the selector and context are correct`);
+
+  const message = `The ${options.selector} node is not found, please confirm whether the selector and context are correct`;
+  console.warn(message);
+
+  return Promise.reject(message);
 }