+++ src/packages/common/functional-component/index.d.ts
@@ -1,7 +1,7 @@
 export declare function selectComponent(context: any, selector: string | Function): any;
 export declare function showFunctionalComponent(options: {
   context?: any;
-  selector: string;
+  selector: string | ((context: any) => any);
   showFunction?: string;
   [k: string]: any;
 }): Promise<unknown>;