+++ src/packages/press-toast/default-options.d.ts
@@ -1,5 +1,5 @@
 export declare const DEFAULT_OPTIONS: {
-  selector: string;
+  selector: string | ((context: any) => any);
   show: boolean;
   mask: boolean;
   message: string;